api/requirements.txt
python=3.11.14
fastapi
fastapi[standard]
uvicorn
requests
databricks.sdk
dotenv
PyJWT
pymongo